PostgreSQL 10 compliant
Get rid of SRF (set-returning functions) appearing in SELECT list.

IMPORTANT: Those LEFT JOIN LATERAL ON [bool] may fool the query planner, causing a wrong plan to be cached if the function is applied to a table with a large number of geometries of one type (eg. POLYGON) and a few of a different one (eg. LINESTRING).
This must be fixed before final 1.2.0 release!